move logic for displaying frame away
